集成建群

+++ backstageProject/src/main/java/com/BYL/lotteryTools/common/service/impl/UploadfileServiceImpl.java
@@ -179,26 +179,30 @@ public class UploadfileServiceImpl implements UploadfileService {
 			 
 			 for (Uploadfile uploadfile : uploadfiles) 
 			 {
-				 	savePath = savePath +uploadfile.getUploadfilepath();
-				 	//2.删除附件
-			 		dirFile = new File(savePath+uploadfile.getUploadRealName());//uploadfile.getUploadRealName()
-			 		logger.info("待删除文件路径："+dirFile);
-			        // 如果dir对应的文件不存在，或者不是一个目录，则退出
-		        	deleteFlag = dirFile.delete();
-		        	if(deleteFlag)
-		        	{//删除附件(清空附件关联newsUuid)
-		        		logger.info("deleteImg==删除原附件文件数据--附件id="+uploadfile.getId()+"--操作人="+LoginUtils.getAuthenticatedUserId(httpSession));
-		        	}
-		        	else
-		        	{
-		        		logger.error("deleteImg ERROR==没有找到要删除的附件文件或删除失败，附件路径为="+savePath+";File.exists="+dirFile.exists());
-		        	}
-				    //删除附件e
-			   		 uploadfile.setModify(uploadfile.getNewsUuid());//放置附件关联uuid
-			   		 uploadfile.setModifyTime(new Timestamp(System.currentTimeMillis()));
-			   		 uploadfile.setIsDeleted(Constants.IS_DELETED);//删除标记
-			   		 uploadfile.setDeleteServiceFile(deleteFlag);//存储是否成功删除文件标记
-			   		 this.update(uploadfile);
+				 if(!"0".equals(uploadfile.getNewsUuid()))
+				 {
+					 savePath = savePath +uploadfile.getUploadfilepath();
+					 	//2.删除附件
+				 		dirFile = new File(savePath+uploadfile.getUploadRealName());//uploadfile.getUploadRealName()
+				 		logger.info("待删除文件路径："+dirFile);
+				        // 如果dir对应的文件不存在，或者不是一个目录，则退出
+			        	deleteFlag = dirFile.delete();
+			        	if(deleteFlag)
+			        	{//删除附件(清空附件关联newsUuid)
+			        		logger.info("deleteImg==删除原附件文件数据--附件id="+uploadfile.getId()+"--操作人="+LoginUtils.getAuthenticatedUserId(httpSession));
+			        	}
+			        	else
+			        	{
+			        		logger.error("deleteImg ERROR==没有找到要删除的附件文件或删除失败，附件路径为="+savePath+";File.exists="+dirFile.exists());
+			        	}
+					    //删除附件e
+				   		 uploadfile.setModify(uploadfile.getNewsUuid());//放置附件关联uuid
+				   		 uploadfile.setModifyTime(new Timestamp(System.currentTimeMillis()));
+				   		 uploadfile.setIsDeleted(Constants.IS_DELETED);//删除标记
+				   		 uploadfile.setDeleteServiceFile(deleteFlag);//存储是否成功删除文件标记
+				   		 this.update(uploadfile);
+				 }
+				 	
 			 }
